+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/tasks/WriteSemverTask.kt
@@ -58,7 +58,8 @@ constructor(
                     |$semver
                     |$semverWithPrefix
                     |
-                """.trimMargin()
+                """
+                    .trimMargin()
             )
         }
     }