+++ semver-gradle-plugin/test/kotlin/com/javiersc/semver/gradle/plugin/setup/Sandbox.kt
@@ -70,7 +70,8 @@ internal fun File.createGitIgnore() {
                 |local.properties
                 |environment/github.env
                 |
-            """.trimMargin()
+            """
+                .trimMargin()
         )
     }
 }