+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 ### Updated
 
+- `gradle -> 7.5.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.29`
-- `gradle -> 7.5`
 
 ## [0.3.0-alpha.4] - 2022-07-10