+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### Updated
 
+- `org.eclipse.jgit:org.eclipse.jgit -> 6.3.0.202209071007-r`
 - `gradle -> 7.5.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.29`