+++ CHANGELOG.md
@@ -17,9 +17,9 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.35`
 - `org.eclipse.jgit:org.eclipse.jgit -> 6.3.0.202209071007-r`
 - `gradle -> 7.5.1`
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.29`
 
 ## [0.3.0-alpha.4] - 2022-07-10