+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.35`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.37`
 - `org.eclipse.jgit:org.eclipse.jgit -> 6.3.0.202209071007-r`
 - `gradle -> 7.5.1`